fix(CustomNodemanager): AddReferences method returns too early
The AddRefrences method loops through a list of (external) references and adds them to the designated objects. When obtaining node handles it also checks that the node belongs to one of the namspaces for the  current node manager. The problem was that it skipped the rest of  the nodes in the list as soon as it found one node that did not belong to the currennt node manager.