Fix argTypes null pointer condition (#1055)
I believe this was a typo:

```
if (argTypes != null || argTypes.Length == 1)
```

(If argTypes was null, then the first part would be false, and the second part of throw an exception. If argsType wasn't null, then the first part would be true and the second part wouldn't be evaluated.) I'm not sure whether the length check should be `== 1` or `>= 1`, nor do I know anything about how this is used, so apologies if I've misunderstood something.