Add minThreadCount and maxThreadCount management
In RequestQueue class constructor the threadpool configuration has been added
in order to handle the minThreadCount and maxThreadCount parameters.
Those parameters are the ones coming from the server config .xml
(MinRequestThreadCount, MaxRequestThreadCount) and at the moment
are not handled in the code.

This modification allows correct setting of the threadpool so that the server can work
with the specified parameters, thus improving performances in a multiclient environment.

Co-Authored-By: macheronte <79026457+macheronte@users.noreply.github.com>