Recover binary decoder from decoding error of custom complex type (#1974)
on a decoding error of an ExtensionObject (custom complex type), the binary decoder resets the read pointer in the stream and decodes the object as ByteArray assuming the underlying complex type interpretation is invalid.
With this method decoding of large chunks can still succeed, see #1705.
The root cause for this issue is currently that client and server may have IOP issues on how structured complex types are encoded. Such warnings must be investigated and fixed.