GDS: manage GDS Server users in Database and make it possible to avoid hardcoded passwords (#2402)
Manage the GDS Server Users in a Database. The implementation is based on the IUsersDatabase Interface which allows the User of the library to provide its own implementaion.
For testing a Json/Linq Users Database implementation is provided.
Passwords are hashed with the PBKDF2 with 10k iterations.

In the default config the standard users with the demo passwords are created, so the behaviour for existing applications is not changed.