Allow bad status code for optional Attributes in ReadNodes (#2730)
ReadNodes() (and some other methods) allow bad status codes for optional values. However, this list is not complete, which means that some optional attributes with a Bad StatusCode cause the read to fail. This Pull Request does not fix something, which is wrong (according to the specification), but rather increases the resilience against servers with problems.

List of optional attributes: https://reference.opcfoundation.org/Core/Part3/v105/docs/5.9